requirements.txt
pandas==2.2.3
numpy==2.1.3
faker==33.3.1
pyyaml==6.0.2
python-dotenv==1.0.1
pydantic==2.10.6
pyarrow==18.1.0
rich==13.9.4
pytest==8.3.4